apply testing fixes, rename fields, and add Railway deployment config
- Backend: validate sgrp_code (5-char alpha), mgrp_code (alpha, max 9)
- Backend: attribute names uppercased, auto-sequence (10,20,30+), duplicate errors
- Backend: sap_item_id validated as numeric max 10 digits
- Backend: new short/long name logic (SAP desc + attrs)
- Backend: settings use env vars (DATABASE_URL, SECRET_KEY, DEBUG)
- Backend: add whitenoise + dj-database-url for Railway
- Backend: railway.toml runs migrate + collectstatic on deploy
- Frontend: materials page - Draft/Final tabs, SAP Material Number/Description
- Frontend: hide local_item_id, wrap long name, fix search
- Frontend: search page - Rank→Hits, sessionStorage persistence
- Frontend: requests page - fix search crash, add project_code filter
- Frontend: railway.toml created for Next.js service
- Root: .gitignore created

+++ meil_backend/itemmaster/models.py
@@ -46,7 +46,7 @@ from Employee.models import Employee
 
 class ItemMaster(models.Model):
     local_item_id = models.AutoField(primary_key=True)
-    sap_item_id = models.IntegerField(null=True, blank=True)
+    sap_item_id = models.BigIntegerField(null=True, blank=True)
 
     mat_type_code = models.ForeignKey(
         "MaterialType.MaterialType",
